split out empty search rec to its own file
