+++ client/reader/search-stream/index.jsx
@@ -126,7 +126,7 @@ class SearchStream extends Component {
 	}
 
 	render() {
-		const { query, suggestions } = this.props;
+		const { query, suggestions, store } = this.props;
 		const emptyContent = <EmptyContent query={ query } />;
 		const sortOrder = this.props.postsStore && this.props.postsStore.sortOrder;
 
@@ -162,6 +162,7 @@ class SearchStream extends Component {
 		return (
 			<Stream
 				{ ...this.props }
+				store={ store }
 				followSource={ SEARCH_RESULTS }
 				listName={ this.props.translate( 'Search' ) }
 				emptyContent={ emptyContent }
